Add X11 abilities to Kitematic
This patch will add a new setting variable in the settings view to set a
`DISPLAY` path for lunching of X11 containers such as `jess/chrome`.

On unix (untested) you simply fill out the input with "unix:0" (if your
current display is 0) and start a X11 container. On Mac OSX you need
start a `socat` daemon. Here's an example:

    socat TCP-LISTEN:6000,reuseaddr,fork UNIX-CLIENT:\"$DISPLAY\"

With a daemon running you can set the X11 display path to the virtualbox
network interface that is running docker ("192.168.59.3").

When lunching a container we'll check for the settings file and inject
into the container options if it exists.

Try running Chrome with `jess/chrome`!

Fixes #270.