Fall-back to image name as key when there is no id
When a container is first started and the image is downloading there is
no id set. With this patch we'll check if the container has a id and
and is downloading, if so we'll substitute it with the image name and
render such.

This will fix the react key props warning in the developer console as
well as maybe give us some sort of performance boost(?)

Signed-off-by: Kristján Oddsson <koddsson@gmail.com>