This change allows to define Volumes where the local folder is not limited to the Users directory but to any folder mounted in the VM.
It recovers from the VBoxManage the mounted folders and it checks the final folder agains this values, and if it matches it allows to specify the path.

This code is related to Issue #1192 opened as a question.

Please, if you have any question, only let me know.

Signed-off-by: Alexandre Vazquez <alexandre.vazquez@gmail.com>

Updating the PR with the changes comments made by @FrenchBen.

Changes after applying the eslint package in Atom.

Removing the homeDir variable because with the new code is not used.

Fixed problem when you tried to remove an added volume.